--- dryad-proto/src/main/proto/Dryad.proto
@@ -1,41 +1,0 @@
-/*
- * Copyright 2022 Korandoru Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-syntax = "proto3";
-option java_package = "io.korandoru.dryad.proto";
-option java_multiple_files = true;
-option java_outer_classname = 'DryadProtos';
-option java_generate_equals_and_hash = true;
-
-package dryad;
-
-message GetRequest {
-  bytes key = 1;
-}
-
-message GetResponse {
-  bool found = 1;
-  bytes key = 2;
-  bytes value = 3;
-}
-
-message PutRequest {
-  bytes key = 1;
-  bytes value = 2;
-}
-
-message PutResponse {
-}
