+++ pom.xml
@@ -76,9 +76,8 @@ limitations under the License.
 
     <modules>
         <module>dryad-client</module>
-        <module>dryad-config</module>
-        <module>dryad-ctl</module>
-        <module>dryad-proto</module>
+        <module>dryad-command</module>
+        <module>dryad-core</module>
         <module>dryad-server</module>
     </modules>
